requirements.txt
flask==2.3.2
mysql-connector-python==8.0.33
paho-mqtt==1.7.1
